requirements.txt
astroid==2.4.1
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django==2.0.12
django-cors-headers==3.1.1
django-enumfields==1.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.5.2
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.6
wrapt==1.12.1
django-test-plus==1.1.1
factory-boy==2.10.0
gunicorn==19.9.0